--- datasophon-api/src/main/test/com/datasophon/api/test/ProcessUtilsTest.java
@@ -1,18 +1,0 @@
-package com.datasophon.api.test;
-
-import com.datasophon.api.utils.ProcessUtils;
-import org.junit.Test;
-
-import java.util.HashMap;
-import java.util.Map;
-
-public class ProcessUtilsTest {
-
-    @Test
-    public void testGenerateGlobalVariables(){
-        Map<String,String> globalVariables =new HashMap<>();
-        String variableName = "test";
-        String variableValue = "test";
-        ProcessUtils.generateClusterVariable(globalVariables,22,variableName,variableValue);
-    }
-}
